Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

It's a good idea to make the calendar use Layout text, though it is not as lightwheight, which is probably the reason it used drawText in the first place.
Considering BiDi related issues in Android, I do believe that a simple BiDi algorithm should be added to drawText to enable BiDi on non BiDi aware programs (such as gmail, facebook, etc..) but left in the StaticLayout to enable control over BiDi aspects for programs that use it, and not to break compatability.

Patch-set: 2
Reviewer: David Kohen <1003738@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1