Update patch set 8
Patch Set 8:

Both the -2 and the +1 came from the original code, before the change. Basically, this creates a 1 pixel border around the text, but, like I said, the proper place to direct this question is the original author of the code. I just maintained the semantics of the original.

Patch-set: 8