Update patch set 1
Patch Set 1:

As a quick note going forward: changes are easier to manage when each different aspect lives in a separate commit, so that it's possible to manage situations like this one where not all changes can be accepted, or when the different aspects being touched are maintained by different people.

Since the relevant parts of this change were manually cherry-picked into Google's master tree, I don't expect any further activity here and I expect that we'll abandon this change in the next few weeks.

Patch-set: 1