Update patch set 2
Patch Set 2:

I've started a thread on the subject on android-platform. At issue here are two independent problems.

One is that there is no BiDi when using Canvas.drawText, and whether that is a good idea or not. The second is that StaticLayout is ill suited for the role the docs give it, which is what drives everyone to use Canvas.drawText in the first place.

The second problem is unrelated to BiDi, except that BiDi languages suffer the most from its manifestation.

Patch-set: 2